test: add test for canonical pcie
